backend/requirements.txt
# FastAPI and Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6

# LangChain and Agent Framework
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
langchain-postgres>=0.0.12
langchain-text-splitters>=0.3.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
pgvector>=0.3.5

# Vector Embeddings and RAG
sentence-transformers>=3.3.0
pypdf>=5.1.0

# Caching
redis>=5.0.0

# Security
cryptography>=41.0.0
pyjwt>=2.8.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0

# Utilities
python-dotenv>=1.0.0
